[18927] Ikke alle ignored poster vises på et job - Implementering af countItems i JobStoreProxy. Tilpasing af countJobs til at returnere long frem for int
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@88727 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ commons/utils/job-store-service-connector/src/main/java/dk/dbc/dataio/commons/utils/jobstore/JobStoreServiceConnector.java
@@ -175,7 +175,7 @@ public class JobStoreServiceConnector {
      * @throws NullPointerException when given null-valued criteria argument
      * @throws JobStoreServiceConnectorException on general failure to produce jobs listing
      */
-    public int countJobs(JobListCriteria criteria) throws NullPointerException, JobStoreServiceConnectorException {
+    public long countJobs(JobListCriteria criteria) throws NullPointerException, JobStoreServiceConnectorException {
         log.trace("JobStoreServiceConnector: listJobs();");
         final StopWatch stopWatch = new StopWatch();
         try {
